+++ LineApplet.java
@@ -2,12 +2,12 @@ import java.awt.Color;
 
 public class LineApplet extends AppletComponent {
 
-    public double x1, y1, x2, y2;
+    public float x1, y1, x2, y2;
     public Color color;
-    public double strokeWeight;
+    public float strokeWeight;
 
-    public LineApplet(double x1, double y1, double x2, double y2, Color color, double strokeWeight, double rotate,
-            Point translate, double scale) {
+    public LineApplet(float x1, float y1, float x2, float y2, Color color, float strokeWeight, float rotate,
+            Point translate, float scale) {
         this.x1 = x1;
         this.y1 = y1;
         this.x2 = x2;
